LayoutTests/imported/w3c/web-platform-tests/digital-credentials/default-permissions-policy.https.sub-expected.txt



PASS Permissions-Policy is by default 'self' for get().
FAIL Permissions-Policy is by default 'self' for create(). promise_rejects_js: function "function() { throw e; }" threw object "NotSupportedError: No credential type was specified." ("NotSupportedError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
